+++ module-item-service/src/main/java/com/fashionmall/item/entity/ItemDiscount.java
@@ -1,6 +1,8 @@
 package com.fashionmall.item.entity;
 
 import com.fashionmall.common.entity.BaseEntity;
+import com.fashionmall.item.enums.ItemDiscountTypeEnum;
+import com.fashionmall.item.enums.ItemStatusEnum;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
